move src/scrolling.ts into src/content
