+++ src/excmds.ts
@@ -114,7 +114,7 @@ import * as SELF from "./.excmds_content.generated"
 Messaging.addListener("excmd_content", Messaging.attributeCaller(SELF))
 import * as DOM from "./dom"
 import { executeWithoutCommandLine } from "@src/content/commandline_content"
-import * as scrolling from "./scrolling"
+import * as scrolling from "@src/content/scrolling"
 // }
 
 //#background_helper
